+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.2.0 | 2017.09.08
+- add file format (use [prettydiff](http://prettydiff.com/))
+- next: add range format
+
 ## 0.1.0 | 2017.09.07
 - add snippets